Migrate to swarm mode initial
