+++ servers/spawners/__init__.py
@@ -1,0 +1,2 @@
+from .plain_docker import DockerSpawner
+from .swarm_mode import SwarmModeSpawner
