--- servers/tests/fake_docker_api_client/__init__.py
@@ -1,4 +1,0 @@
-"""
-This is directly taken from docker-py 2.0.2 with some modifications
-License is here https://github.com/docker/docker-py/blob/master/LICENSE
-"""
